# bpo-36793: Remove unneeded __str__ definitions.