--- a/Lib/email/charset.py
+++ b/Lib/email/charset.py
@@ -241,10 +241,8 @@ class Charset:
         self.output_codec = CODEC_MAP.get(self.output_charset,
                                           self.output_charset)
 
-    def __str__(self):
+    def __repr__(self):
         return self.input_charset.lower()
-
-    __repr__ = __str__
 
     def __eq__(self, other):
         return str(self) == str(other).lower()