--- a/Lib/sre_constants.py
+++ b/Lib/sre_constants.py
@@ -59,10 +59,8 @@ class _NamedIntConstant(int):
         self.name = name
         return self
 
-    def __str__(self):
+    def __repr__(self):
         return self.name
-
-    __repr__ = __str__
 
 MAXREPEAT = _NamedIntConstant(MAXREPEAT, 'MAXREPEAT')