Classes that define `__str__` the same as `__repr__` can
just inherit it from object.


<!-- issue-number: [bpo-36793](https://bugs.python.org/issue36793) -->
https://bugs.python.org/issue36793
<!-- /issue-number -->