@skrah 2019-05-04
Hm, what is the benefit of this? I find the existing code much clearer.