@serhiy-storchaka 2019-05-04
The benefit is consistency. When you define `__repr__` in a subclass of `list` or `dict` or in any user class, this will affect the result of `str()`. But `int`, `float` and `complex` are exceptional. There were reasons to be exceptional in Python 2, but this is no longer needed in Python 3.

Similarly, we usually do not implement `__ne__` if define `__eq__`, because the default implementation of `__ne__` falls back to `__eq__`.