@skrah 2019-05-04
But falling back is a bit slower, right (I did not look at the relevant code)?