@skrah 2019-05-04
@serhiy-storchaka Thanks, I have no objections then.  Probably people do expect `__repr__` to behave like you said.

Personally, I still always define both `__eq__`and `__ne__`. :)